Windows inline ps1 when exec privilege restricted
In restricted windows machines, it is forbidden
to specify a powershell execution policy and
it is usually forbidden to execute an unsigned script.

Inlining the script as an encoded command avoids that problem.
Note that the encoded command cannot be passed a parameter:
so we need to place the value of the image path inside the script
and generate it dynamically on every invokation.